pyproject.toml
[project]
name = "text-to-sql-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chainlit>=2.9.6",
    "chromadb>=1.5.1",
    "fastapi>=0.133.0",
    "langchain>=1.2.10",
    "langchain-ollama>=1.0.1",
    "loguru>=0.7.0",
    "psycopg2-binary>=2.9.9",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-dotenv>=1.0.0",
    "sentence-transformers>=5.2.3",
    "sqlalchemy>=2.0.47",
    "sqlparse>=0.5.5",
    "streamlit>=1.54.0",
    "uvicorn>=0.41.0",
]

[tool.setuptools.packages.find]
include = ["src*"]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-mock>=3.15.1",
]
